another

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ float floatType;
 char charType;
 long int longType;
 long long int longLongType;
-printf("size of an int: %zu byte(s)\n", sizeof(intType));
-printf("size of a float: %zu byte(s)\n", sizeof(floatType));
-printf("size of a char: %zu byte(s)\n", sizeof(charType));
-printf("size of a long int: %zu byte(s)\n", sizeof(longType));
-printf("size of a long long int: %zu byte(s)\n", sizeof(longLongType));
+printf("size of an int: %lu byte(s)\n", (unsigned long)sizeof(intType));
+printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(floatType));
+printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(charType));
+printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(longType));
+printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(longLongType));
 return (0);
 }
